- website - merge file report.js into browser.js.
