+++ browser.js
@@ -7,13 +7,16 @@
 */
 
 /*property
-    checked, create, disable, display, error, focus, forEach, function,
-    getElementById, innerHTML, join, length, map, onchange, onclick, onscroll,
-    property, querySelectorAll, scrollTop, select, split, style, title, value
+    checked, closure, column, context, create, disable, display, edition,
+    exports, filter, focus, forEach, froms, fudge, functions, getElementById,
+    global, id, innerHTML, isArray, join, json, keys, length, level, line,
+    lines, map, message, module, name, names, onchange, onclick, onscroll,
+    option, parameters, parent, property, push, querySelectorAll, replace, role,
+    scrollTop, select, signature, sort, split, stop, style, title, value,
+    warnings
 */
 
 import jslint from "./jslint.js";
-import report from "./report.js";
 
 // This is the web script companion file for JSLint. It includes code for
 // interacting with the browser and displaying the reports.
@@ -23,50 +26,261 @@ let rx_separator = /[\s,;'"]+/;
 
 let fudge_unit = 0;
 
-const aux = document.getElementById("JSLINT_AUX");
-const boxes = document.querySelectorAll("[type=checkbox]");
-const fudge = document.getElementById("JSLINT_FUDGE");
-const global = document.getElementById("JSLINT_GLOBAL");
-const number = document.getElementById("JSLINT_NUMBER");
-const property = document.getElementById("JSLINT_PROPERTY");
-const property_fieldset = document.getElementById("JSLINT_PROPERTYFIELDSET");
-const report_field = document.getElementById("JSLINT_REPORT");
-const report_list = document.getElementById("JSLINT_REPORT_LIST");
-const source = document.getElementById("JSLINT_SOURCE");
-const select = document.getElementById("JSLINT_SELECT");
-const warnings = document.getElementById("JSLINT_WARNINGS");
-const warnings_list = document.getElementById("JSLINT_WARNINGS_LIST");
+const elem_aux = document.getElementById("JSLINT_AUX");
+const elem_boxes = document.querySelectorAll("[type=checkbox]");
+const elem_fudge = document.getElementById("JSLINT_FUDGE");
+const elem_global = document.getElementById("JSLINT_GLOBAL");
+const elem_number = document.getElementById("JSLINT_NUMBER");
+const elem_property = document.getElementById("JSLINT_PROPERTY");
+const elem_property_fieldset = document.getElementById(
+    "JSLINT_PROPERTYFIELDSET"
+);
+const elem_report_field = document.getElementById("JSLINT_REPORT");
+const elem_report_list = document.getElementById("JSLINT_REPORT_LIST");
+const rx_amp = /&/g;
+const rx_gt = />/g;
+const rx_lt = /</g;
+const elem_select = document.getElementById("JSLINT_SELECT");
+const elem_source = document.getElementById("JSLINT_SOURCE");
+const elem_warnings = document.getElementById("JSLINT_WARNINGS");
+const elem_warnings_list = document.getElementById("JSLINT_WARNINGS_LIST");
+
+function entityify(string) {
+
+// Replace & < > with less destructive entities.
+
+    return String(
+        string
+    ).replace(
+        rx_amp,
+        "&amp;"
+    ).replace(
+        rx_lt,
+        "&lt;"
+    ).replace(
+        rx_gt,
+        "&gt;"
+    );
+}
+
+function error_report(data) {
+
+// Produce the HTML Error Report.
+
+// <cite><address>LINE_NUMBER</address>MESSAGE</cite>
+// <samp>EVIDENCE</samp>
+
+    let fudge = Number(Boolean(data.option.fudge));
+    let output = [];
+    if (data.stop) {
+        output.push("<center>JSLint was unable to finish.</center>");
+    }
+    data.warnings.forEach(function (warning) {
+        output.push(
+            "<cite><address>",
+            entityify(warning.line + fudge),
+            ".",
+            entityify(warning.column + fudge),
+            "</address>",
+            entityify(warning.message),
+            "</cite><samp>",
+            entityify(data.lines[warning.line] || ""),
+            "</samp>"
+        );
+    });
+    return output.join("");
+}
+
+function function_report(data) {
+
+// Produce the HTML Function Report.
+
+// <dl class=LEVEL><address>LINE_NUMBER</address>FUNCTION_NAME_AND_SIGNATURE
+//     <dt>DETAIL</dt><dd>NAMES</dd>
+// </dl>
+
+    let fudge = Number(Boolean(data.option.fudge));
+    let mode = (
+        data.module
+        ? "module"
+        : "global"
+    );
+    let output = [];
+
+    if (data.json) {
+        return (
+            data.warnings.length === 0
+            ? "<center>JSON: good.</center>"
+            : "<center>JSON: bad.</center>"
+        );
+    }
+
+    function detail(title, array) {
+        if (Array.isArray(array) && array.length > 0) {
+            output.push(
+                "<dt>",
+                entityify(title),
+                "</dt><dd>",
+                array.join(", "),
+                "</dd>"
+            );
+        }
+    }
+
+    if (data.functions.length === 0) {
+        output.push("<center>There are no functions.</center>");
+    }
+    let global = Object.keys(data.global.context).sort();
+    let froms = data.froms.sort();
+    let exports = Object.keys(data.exports).sort();
+    if (global.length + froms.length + exports.length > 0) {
+        output.push("<dl class=level0>");
+        detail(mode, global);
+        detail("import from", froms);
+        detail("export", exports);
+        output.push("</dl>");
+    }
+
+    if (data.functions.length > 0) {
+        data.functions.forEach(function (the_function) {
+            let context = the_function.context;
+            let list = Object.keys(context);
+            output.push(
+                "<dl class=level",
+                entityify(the_function.level),
+                "><address>",
+                entityify(the_function.line + fudge),
+                "</address><dfn>",
+                (
+                    the_function.name === "=>"
+                    ? entityify(the_function.signature) + " =>"
+                    : (
+                        typeof the_function.name === "string"
+                        ? "<b>«" + entityify(the_function.name) + "»</b>"
+                        : "<b>" + entityify(the_function.name.id) + "</b>"
+                    )
+                ) + entityify(the_function.signature),
+                "</dfn>"
+            );
+            if (Array.isArray(the_function.parameters)) {
+                let params = [];
+                the_function.parameters.forEach(function extract(name) {
+                    if (name.id === "{" || name.id === "[") {
+                        name.names.forEach(extract);
+                    } else {
+                        if (name.id !== "ignore") {
+                            params.push(name.id);
+                        }
+                    }
+                });
+                detail(
+                    "parameter",
+                    params.sort()
+                );
+            }
+            list.sort();
+            detail("variable", list.filter(function (id) {
+                let the_variable = context[id];
+                return (
+                    the_variable.role === "variable"
+                    && the_variable.parent === the_function
+                );
+            }));
+            detail("exception", list.filter(function (id) {
+                return context[id].role === "exception";
+            }));
+            detail("closure", list.filter(function (id) {
+                let the_variable = context[id];
+                return (
+                    the_variable.closure === true
+                    && the_variable.parent === the_function
+                );
+            }));
+            detail("outer", list.filter(function (id) {
+                let the_variable = context[id];
+                return (
+                    the_variable.parent !== the_function
+                    && the_variable.parent.id !== "(global)"
+                );
+            }));
+            detail(mode, list.filter(function (id) {
+                return context[id].parent.id === "(global)";
+            }));
+            detail("label", list.filter(function (id) {
+                return context[id].role === "label";
+            }));
+            output.push("</dl>");
+        });
+    }
+    output.push(
+        "<center>JSLint edition ",
+        entityify(data.edition),
+        "</center>"
+    );
+    return output.join("");
+}
+
+function property_directive(data) {
+
+// Produce the /*property*/ directive.
+
+    let not_first = false;
+    let output = ["/*property"];
+    let length = 1111;
+    let properties = Object.keys(data.property);
+
+    if (properties.length > 0) {
+        properties.sort().forEach(function (key) {
+            if (not_first) {
+                output.push(",");
+                length += 2;
+            }
+            not_first = true;
+            if (length + key.length >= 80) {
+                length = 4;
+                output.push("\n   ");
+            }
+            output.push(" ", key);
+            length += key.length;
+        });
+        output.push("\n*/\n");
+        return output.join("");
+    }
+}
 
 function show_numbers() {
-    number.value = source.value.split(rx_crlf).map(function (ignore, index) {
+    elem_number.value = elem_source.value.split(rx_crlf).map(function (
+        ignore,
+        index
+    ) {
         return index + fudge_unit;
     }).join("\n");
 }
 
 function clear() {
-    aux.style.display = "none";
-    number.value = "";
-    property.value = "";
-    property_fieldset.style.display = "none";
-    report_field.style.display = "none";
-    report_list.innerHTML = "";
-    source.focus();
-    source.value = "";
-    warnings.style.display = "none";
-    warnings_list.innerHTML = "";
+    elem_aux.style.display = "none";
+    elem_number.value = "";
+    elem_property.value = "";
+    elem_property_fieldset.style.display = "none";
+    elem_report_field.style.display = "none";
+    elem_report_list.innerHTML = "";
+    elem_source.focus();
+    elem_source.value = "";
+    elem_warnings.style.display = "none";
+    elem_warnings_list.innerHTML = "";
 }
 
 function fudge_change() {
-    fudge_unit = Number(fudge.checked);
+    fudge_unit = Number(elem_fudge.checked);
     show_numbers();
 }
 
 function clear_options() {
-    boxes.forEach(function (node) {
+    elem_boxes.forEach(function (node) {
         node.checked = false;
     });
     fudge_change();
-    global.innerHTML = "";
+    elem_global.innerHTML = "";
 }
 
 function call_jslint() {
@@ -74,7 +288,7 @@ function call_jslint() {
 // First build the option object.
 
     let option = Object.create(null);
-    boxes.forEach(function (node) {
+    elem_boxes.forEach(function (node) {
         if (node.checked) {
             option[node.title] = true;
         }
@@ -82,9 +296,9 @@ function call_jslint() {
 
 // Call JSLint with the source text, the options, and the predefined globals.
 
-    let global_string = global.value;
+    let global_string = elem_global.value;
     let result = jslint(
-        source.value,
+        elem_source.value,
         option,
         (
             global_string === ""
@@ -95,47 +309,47 @@ function call_jslint() {
 
 // Generate the reports.
 
-    let error_html = report.error(result);
-    let function_html = report.function(result);
-    let property_text = report.property(result);
+    let error_html = error_report(result);
+    let function_html = function_report(result);
+    let property_text = property_directive(result);
 
 // Display the reports.
 
-    warnings_list.innerHTML = error_html;
-    warnings.style.display = (
+    elem_warnings_list.innerHTML = error_html;
+    elem_warnings.style.display = (
         error_html.length === 0
         ? "none"
         : "block"
     );
 
-    report_list.innerHTML = function_html;
-    report_field.style.display = "block";
+    elem_report_list.innerHTML = function_html;
+    elem_report_field.style.display = "block";
     if (property_text) {
-        property.value = property_text;
-        property_fieldset.style.display = "block";
-        property.scrollTop = 0;
-        select.disable = false;
+        elem_property.value = property_text;
+        elem_property_fieldset.style.display = "block";
+        elem_property.scrollTop = 0;
+        elem_select.disable = false;
     } else {
-        property_fieldset.style.display = "none";
-        select.disable = true;
+        elem_property_fieldset.style.display = "none";
+        elem_select.disable = true;
     }
-    aux.style.display = "block";
-    source.select();
+    elem_aux.style.display = "block";
+    elem_source.select();
 }
 
-fudge.onchange = fudge_change;
+elem_fudge.onchange = fudge_change;
 
-source.onchange = function (ignore) {
+elem_source.onchange = function (ignore) {
     show_numbers();
 };
 
-source.onscroll = function () {
-    let ss = source.scrollTop;
-    number.scrollTop = ss;
-    let sn = number.scrollTop;
+elem_source.onscroll = function () {
+    let ss = elem_source.scrollTop;
+    elem_number.scrollTop = ss;
+    let sn = elem_number.scrollTop;
     if (ss > sn) {
         show_numbers();
-        number.scrollTop = ss;
+        elem_number.scrollTop = ss;
     }
 };
 
@@ -148,11 +362,11 @@ document.querySelectorAll("[name='clear']").forEach(function (node) {
 });
 
 document.getElementById("JSLINT_SELECT").onclick = function () {
-    property.focus();
-    property.select();
+    elem_property.focus();
+    elem_property.select();
 };
 document.getElementById("JSLINT_CLEAR_OPTIONS").onclick = clear_options;
 
 fudge_change();
-source.select();
-source.focus();
+elem_source.select();
+elem_source.focus();
